backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
channels>=4.0.0
daphne>=4.0.0
django-filter>=23.2
drf-spectacular>=0.26.0
djoser>=2.2.0
djangorestframework-simplejwt>=5.3.1
